+++ R/set_opts.R
@@ -21,15 +21,22 @@
 #'    freedom, matching the SUDAAN calculation. NHIS users, be sure to also type
 #'    `options(surveytable.adjust_svyciprop.df_method = "NHIS")`.
 #'
+#' `output` determines how the output is printed.
+#'
+#' * `"auto"` (default): automatically select the printing method, depending on the
+#' destination (such as screen, HTML, or PDF / LaTeX).
+#' * `"huxtable"`, `"gt"`, or `"kableExtra"`: use this package for printing. Be sure
+#' that this package is installed.
+#' * `"raw"`: unformatted / raw output. This is useful for getting lots of significant digits.
+#'
 #' @param mode `"general"` or `"NCHS"`. See below for details.
 #' @param count round counts to the nearest: integer (`"int"`) or one thousand (`"1k"`)
 #' @param lpe identify low-precision estimates?
 #' @param drop_na drop missing values (`NA`)? Categorical variables only.
 #' @param max_levels a categorical variable can have at most this many levels. Used to avoid printing huge tables.
 #' @param csv     name of a CSV file or `""` to turn off CSV output.
-#' @param output package to use for printing. One of `"huxtable"`, `"gt"`, or `"kableExtra"`.
-#' For the last two, be sure that this package is installed. `"auto"` (default) = automatically
-#' select `huxtable` for screen, `gt` for HTML, or `kableExtra` for PDF (LaTeX).
+#' @param output how the output is printed. `"auto"` (default); `"huxtable"`, `"gt"`, or
+#' `"kableExtra"`; or `"raw"`.
 #'
 #' @return (Nothing.)
 #' @family options
@@ -59,13 +66,15 @@ set_opts = function(
     mode %<>% .mymatch(c("nchs", "general"))
     if (mode == "nchs") {
       message("* Mode: NCHS.")
-      options(surveytable.tx_count = ".tx_count_1k"
+      options(surveytable.do_tx = TRUE
+        , surveytable.tx_count = ".tx_count_1k"
         , surveytable.names_count = c("n", "Number (000)", "SE (000)", "LL (000)", "UL (000)")
         , surveytable.find_lpe = TRUE
         , surveytable.adjust_svyciprop = TRUE)
     } else if (mode == "general") {
       message("* Mode: General.")
-      options(surveytable.tx_count = ".tx_count_int"
+      options(surveytable.do_tx = TRUE
+        , surveytable.tx_count = ".tx_count_int"
         , surveytable.names_count = c("n", "Number", "SE", "LL", "UL")
         , surveytable.find_lpe = FALSE
         , surveytable.adjust_svyciprop = FALSE)
@@ -76,11 +85,13 @@ set_opts = function(
     count %<>% .mymatch(c("int", "1k"))
     if (count == "int") {
       message("* Rounding counts to the nearest integer.")
-      options(surveytable.tx_count = ".tx_count_int"
+      options(surveytable.do_tx = TRUE
+        , surveytable.tx_count = ".tx_count_int"
         , surveytable.names_count = c("n", "Number", "SE", "LL", "UL"))
     } else if (count == "1k") {
       message("* Rounding counts to the nearest thousand.")
-      options(surveytable.tx_count = ".tx_count_1k"
+      options(surveytable.do_tx = TRUE
+        , surveytable.tx_count = ".tx_count_1k"
         , surveytable.names_count = c("n", "Number (000)", "SE (000)", "LL (000)", "UL (000)"))
     }
   }
@@ -127,9 +138,13 @@ set_opts = function(
   }
 
   if (!is.null(output)) {
-    output %<>% .mymatch(c("huxtable", "gt", "kableExtra", "auto"))
+    output %<>% .mymatch(c("huxtable", "gt", "kableExtra", "auto", "raw"))
     if (output == "auto") {
       message("* Printing with huxtable for screen, gt for HTML, or kableExtra for PDF.")
+    } else if (output == "raw") {
+      options(surveytable.do_tx = FALSE
+              , surveytable.names_count = c("n", "Number", "SE", "LL", "UL"))
+      message(glue("* Generating unformatted / raw output."))
     } else {
       message(glue("* Printing with {output}."))
     }
@@ -144,14 +159,20 @@ set_opts = function(
 #' @export
 show_opts = function() {
 
-  tx_count = getOption("surveytable.tx_count")
-  assert_that(is.string(tx_count), nzchar(tx_count))
-  switch(tx_count
-         , ".tx_count_int" = "* Rounding counts to the nearest integer."
-         , ".tx_count_1k" = "* Rounding counts to the nearest thousand."
-         , ".tx_none" = "* Not rounding counts."
-         , " * Count rounding function: " %>% paste0(tx_count)
-         ) %>% message
+  do_tx = getOption("surveytable.do_tx")
+  assert_that(do_tx %in% c(TRUE, FALSE))
+  if (do_tx == FALSE) {
+    message("* Not rounding.")
+  } else {
+    tx_count = getOption("surveytable.tx_count")
+    assert_that(is.string(tx_count), nzchar(tx_count))
+    switch(tx_count
+           , ".tx_count_int" = "* Rounding counts to the nearest integer."
+           , ".tx_count_1k" = "* Rounding counts to the nearest thousand."
+           , ".tx_none" = "* Not rounding counts."
+           , " * Count rounding function: " %>% paste0(tx_count)
+    ) %>% message
+  }
 
   lpe = getOption("surveytable.find_lpe")
   assert_that(is.flag(lpe), lpe %in% c(TRUE, FALSE))
@@ -201,6 +222,7 @@ show_opts = function() {
          , ".print_gt" = "* Printing with gt."
          , ".print_kableextra" = "* Printing with kableExtra."
          , ".print_auto" = "* Printing with huxtable for screen, gt for HTML, or kableExtra for PDF."
+         , ".print_raw" = "* Generating unformatted / raw output."
          , glue("Printing with a custom function: {xx}")) %>% message
   invisible(NULL)
 }