+++ R/tab.R
@@ -193,7 +193,10 @@ tab = function(...
 	              , all(pco$has.flag %in% names(pco$descriptions)))
 	}
 
-	mmc = getOption("surveytable.tx_count") %>% do.call(list(mmcr[,c("x", "s", "ll", "ul")]))
+  mmc = mmcr[,c("x", "s", "ll", "ul")]
+  if (getOption("surveytable.do_tx")) {
+    mmc = getOption("surveytable.tx_count") %>% do.call(list(mmc))
+  }
 	mmc$counts = mmcr$counts
 	mmc = mmc[,c("counts", "x", "s", "ll", "ul")]
 	names(mmc) = getOption("surveytable.names_count")
@@ -242,7 +245,10 @@ tab = function(...
 	              , all(ppo$has.flag %in% names(ppo$descriptions)))
 	}
 
-	mp2 = getOption("surveytable.tx_prct") %>% do.call(list(ret[,c("Proportion", "SE", "LL", "UL")]))
+	mp2 = ret[,c("Proportion", "SE", "LL", "UL")]
+	if (getOption("surveytable.do_tx")) {
+	  mp2 = getOption("surveytable.tx_prct") %>% do.call(list(mp2))
+	}
 	names(mp2) = getOption("surveytable.names_prct")
 
 	##