+++ R/tab_rate.R
@@ -89,7 +89,9 @@ tab_rate = function(vr, pop
   }
   m1 = m1[,cc]
   cc = c("Rate", "SE", "LL", "UL")
-  m1[,cc] = getOption("surveytable.tx_rate") %>% do.call(list(m1[,cc]))
+  if (getOption("surveytable.do_tx")) {
+    m1[,cc] = getOption("surveytable.tx_rate") %>% do.call(list(m1[,cc]))
+  }
 
   attr(m1, "title") = glue("{.getvarname(env$survey, vr)} (rate per {per} population)")
   attr(m1, "num") = 2:6