+++ R/total_rate.R
@@ -39,7 +39,9 @@ total_rate = function(pop
   }
   m1 = m1[,cc]
   cc = c("Rate", "SE", "LL", "UL")
-  m1[,cc] = getOption("surveytable.tx_rate") %>% do.call(list(m1[, cc]))
+  if (getOption("surveytable.do_tx")) {
+    m1[,cc] = getOption("surveytable.tx_rate") %>% do.call(list(m1[,cc]))
+  }
 
   attr(m1, "title") = paste("Total (rate per", per, "population)")
   attr(m1, "num") = 1:5