+++ R/z_tab_numeric.R
@@ -27,10 +27,12 @@
   ret$SD = (svyvar(frm, design, na.rm = TRUE)
             %>% as.numeric %>% sqrt %>% suppressWarnings)
 
-  cc = c("Mean", "SEM", "SD")
-  ret[,cc] = getOption("surveytable.tx_numeric") %>% do.call(list(ret[,cc]))
-  cc = "% known"
-  ret[,cc] = getOption("surveytable.tx_prct") %>% do.call(list(ret[,cc]))
+  if (getOption("surveytable.do_tx")) {
+    cc = c("Mean", "SEM", "SD")
+    ret[,cc] = getOption("surveytable.tx_numeric") %>% do.call(list(ret[,cc]))
+    cc = "% known"
+    ret[,cc] = getOption("surveytable.tx_prct") %>% do.call(list(ret[,cc]))
+  }
 
   ret
 }